handle importXmlInternalError gracefully in RO.php
parseResponse() was throwing "Cannot access offset of type string on string" from line 1301 when response from FlexiBee came with following data:
```
["@version" => "1.0", "success" => "false", "stats" => ["created" => "0", "updated" => "1", "deleted" => "0", "skipped" => "0", "failed" => "0"], "results" => [["errors" => [["message" => "Interní chyba aplikace.", "messageCode" => "importXmlInternalError"]]]]]
```
That made it hard to diagnose problems, because it stripped out relevant information from FlexiBee response. This commit adds fallback for this case when message is not found where expected.